requirements.txt
build==1.4.3
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.2
hctvwrapper==0.1.0
idna==3.11
packaging==26.1
pip-tools==7.5.3
pyproject_hooks==1.2.0
python-dotenv==1.2.2
pythonping==1.1.4
requests==2.33.0
setuptools==82.0.1
urllib3==2.6.3
websockets==16.0
wheel==0.46.3
